Scope Jacinta and Ypsiloid AST accessors to their Ast companions
The array/string/long/is…/object… accessor extensions over the opaque
`Json.Ast` and `Yaml.Ast` types were top-level, so SN-742.1 (rightly) flagged
them as unexported — but they're internal to the AST representation and their
generic names would clash in the `soundness` umbrella (jacinta and ypsiloid
both define `array`, `long`, `locate`, …).

Moving each block into the opaque type's companion `object Ast` keeps them in
implicit scope wherever a `Json.Ast`/`Yaml.Ast` is used (no import needed),
without being top-level or entering the umbrella. The jacinta `arrayLength`/
`objectSize` extensions are folded into the companion (they previously
delegated to identically-named helpers there); the ypsiloid block had no such
clash. One inline `summonFrom` site that resolved `primitive` via the old
package-level extension is qualified to `Yaml.primitive`.
